+++ tests/spec/core_functions/color/blackness.rs
@@ -64,7 +64,6 @@ mod error {
     }
 }
 #[test]
-#[ignore] // wrong result
 fn fraction() {
     assert_eq!(
         runner().ok("@use \'sass:color\';\
@@ -89,7 +88,6 @@ mod middle {
     use super::runner;
 
     #[test]
-    #[ignore] // wrong result
     fn half_whiteness() {
         assert_eq!(
             runner().ok("@use \'sass:color\';\
@@ -100,7 +98,6 @@ mod middle {
         );
     }
     #[test]
-    #[ignore] // wrong result
     fn high_whiteness() {
         assert_eq!(
             runner().ok("@use \'sass:color\';\
@@ -111,7 +108,6 @@ mod middle {
         );
     }
     #[test]
-    #[ignore] // wrong result
     fn zero_whiteness() {
         assert_eq!(
             runner().ok("@use \'sass:color\';\
@@ -133,7 +129,6 @@ fn min() {
     );
 }
 #[test]
-#[ignore] // wrong result
 fn named() {
     assert_eq!(
         runner().ok("@use \'sass:color\';\